cloud-sql-manager/requirements.txt
flask==2.3.3
gunicorn==20.1.0
google-api-python-client==2.15.0
google-auth==2.3.0
